Move the DetailsComponent into the Show namespace
