+++ app/components/show/admin_policy_component.html.erb
@@ -12,7 +12,7 @@
     <%= render OverviewComponent.new(presenter: presenter) %>
   </div>
   <div class="col-md-6">
-    <%= render DetailsComponent.new(presenter: presenter) %>
+    <%= render Show::DetailsComponent.new(presenter: presenter) %>
   </div>
 </div>
