+++ app/components/show/details_component.html.erb
