+++ app/components/show/item_component.html.erb
@@ -23,7 +23,7 @@
     <%= render OverviewComponent.new(presenter: presenter) %>
   </div>
   <div class="col-md-6">
-    <%= render DetailsComponent.new(presenter: presenter) %>
+    <%= render Show::DetailsComponent.new(presenter: presenter) %>
   </div>
 </div>
